Update 3700send
This is a super simple implementation of a window - I will adjust the ACK statement sent by the receiver.
Basically it's set to mirror slow start and use additive increase, multiplicative decrease to change the size (probably not the most efficient
but we should try to test and make sure the program runs at all before thinking about that).
I'm not sure about the while loop I added for sending packets without an ACK but my idea was that we have a current window (estimated) that is equal to the window and every time we send a new packet we decrease the current window until we've run out of packets to send without an ACK. When we receive an ACK the window will either
double or increase by 1. I'm not sure about the timing of this but my thought process was that if a message is receiving while the new packets are being sent
the window will change (but I could be wrong about this because the while loop might just force packets to be sent - i.e. we might have to change the order of these).
Overall we will probably have to do some testing to make this work but this should get us going somewhere.